requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
cycler==0.10.0
decorator==4.1.2
functools32==3.2.3.post2
GDAL==2.1.3
idna==2.6
lxml==4.0.0
matplotlib==2.0.2
netCDF4==1.2.9
nose==1.3.7
numpy==1.13.1
olefile==0.44
opencv-python==3.2.0.8
Pillow==4.2.1
pykml==0.1.3
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
Shapely==1.6.1
six==1.10.0
subprocess32==3.2.7
urllib3==1.22
utm==0.4.2
validators==0.12.0
